+++ assets/scripts/theme/slideshow.ts
@@ -1,9 +1,5 @@
-import UIkit from './UIkit';
-
 const initSlideshows = () => {
-  const slideshows = UIkit.slideshow('.uk-slideshow', {});
-
-  console.log('slideshows loaded', slideshows);
+  console.log('slideshows loaded');
 };
 
 export default initSlideshows;