+++ functions.php
@@ -23,7 +23,7 @@ use MC\App\Fields\FieldGroups\PageBuilderFieldGroup;
  * Define Theme Version
  * Define Theme directories
  */
-define('THEME_VERSION', '3.1.0');
+define('THEME_VERSION', '3.2.0');
 define('MC_THEME_DIR', trailingslashit(get_template_directory()));
 define('MC_THEME_PATH_URL', trailingslashit(get_template_directory_uri()));
 
@@ -33,39 +33,44 @@ require_once MC_THEME_DIR . 'vendor/autoload.php';
 /**
  * Theme Setup
  */
-add_action('after_setup_theme', function () {
+add_action(
+    'after_setup_theme',
+    function () {
+        (new Init())
+            ->add(new Setup())
+            ->add(new Scripts())
+            ->add(new Media())
+            ->add(new Shortcodes())
+            ->add(new ACF())
+            ->add(new Options())
+            ->add(new Modules())
+            ->add(new SiteOptionsFieldGroup())
+            ->add(new PageBuilderFieldGroup())
+            // ->add(new RegisterBlocks())
+            ->initialize();
 
-    (new Init())
-        ->add(new Setup())
-        ->add(new Scripts())
-        ->add(new Media())
-        ->add(new Shortcodes())
-        ->add(new ACF())
-        ->add(new Options())
-        ->add(new Modules())
-        ->add(new SiteOptionsFieldGroup())
-        ->add(new PageBuilderFieldGroup())
-        // ->add(new RegisterBlocks())
-        ->initialize();
+        // Translation setup
+        load_theme_textdomain('mc-starter', MC_THEME_DIR . '/languages');
 
-    // Translation setup
-    load_theme_textdomain('mc-starter', MC_THEME_DIR . '/languages');
+        // Let WordPress manage the document title.
+        add_theme_support('title-tag');
 
-    // Let WordPress manage the document title.
-    add_theme_support('title-tag');
+        // Add automatic feed links in header
+        add_theme_support('automatic-feed-links');
 
-    // Add automatic feed links in header
-    add_theme_support('automatic-feed-links');
+        // Add Post Thumbnail Image sizes and support
+        add_theme_support('post-thumbnails');
 
-    // Add Post Thumbnail Image sizes and support
-    add_theme_support('post-thumbnails');
-
-    // Switch default core markup to output valid HTML5.
-    add_theme_support('html5', [
-        'search-form',
-        'comment-form',
-        'comment-list',
-        'gallery',
-        'caption'
-    ]);
-});
+        // Switch default core markup to output valid HTML5.
+        add_theme_support(
+            'html5',
+            [
+                'search-form',
+                'comment-form',
+                'comment-list',
+                'gallery',
+                'caption'
+            ]
+        );
+    }
+);