+++ vite.config.ts
@@ -1,33 +1,37 @@
 import { defineConfig } from 'vite';
+import vue from '@vitejs/plugin-vue';
 
 export default defineConfig({
   build: {
     rollupOptions: {
       input: {
-        ['theme-scripts']: './assets/scripts/theme.ts',
-        ['admin-scripts']: './assets/scripts/admin.ts',
-        ['theme-styles']: './assets/styles/theme.scss',
-        ['editor-styles']: './assets/styles/editor.scss',
+        ['theme-scripts']: './assets/scripts/theme-scripts.ts',
+        ['admin-scripts']: './assets/scripts/admin-scripts.ts',
+        ['theme-styles']: './assets/styles/theme-styles.scss',
+        ['editor-styles']: './assets/styles/editor-styles.scss',
       },
       output: {
         dir: './build',
-        entryFileNames: 'scripts/[name].min.js',
+        entryFileNames: (asset) => {
+          console.log('entryFileNames', asset.name);
+          return `scripts/[name].min.js`;
+        },
         assetFileNames: (asset) => {
-          const hasStyleExt =
-            asset.name?.includes('.css') || asset.name?.includes('.scss');
-          if (hasStyleExt) {
-            return `styles/[name]-styles.min.[ext]`;
-          }
-          return `[name]-styles.[ext]`;
+          console.log('assetFileNames', asset);
+          return `styles/[name].min.[ext]`;
         },
       },
     },
     sourcemap: true,
   },
-  plugins: [],
-  resolve: {
-    alias: {
-      'uikit-util': 'uikit/src/js/util',
-    },
-  },
+  plugins: [
+    vue({
+      template: {
+        compilerOptions: {
+          // treat all tags with a dash as custom elements
+          isCustomElement: (tag) => tag.includes('-'),
+        },
+      },
+    }),
+  ],
 });